Add Python tests
Compare the python module output to the oracle
