+++ Manylinux2014_Compliant_Source/pkg/build.sh
@@ -17,4 +17,4 @@ cmake -DCMAKE_BUILD_TYPE=Release ../src && make -j"$(nproc)" && make install
 
 cd ..
 cp src/include/utils.h Manylinux2014_Compliant_Source/pkg/IR2Vec/
-cp vocabulary/seedEmbeddingVocab-300-llvm12.txt Manylinux2014_Compliant_Source/pkg/IR2Vec/
+cp vocabulary/seedEmbeddingVocab-*.txt Manylinux2014_Compliant_Source/pkg/IR2Vec/
